feat(presets)!: delete the runtime-generated profile system
Tasks 1-4 made presets fully datapack-driven; UrbexProfile, its
Configuration/ProfileSetup TOML-migration machinery, and the dead
PacketRequestProfile/PacketReturnProfileToClient pair (handle() has
been an empty stub since 1.14) were unreferenced by any live code
path. Delete them along with their superseded density tests
(UrbexProfileDensityTest, ProfileSetupDensityTest).

CommandSaveProfile becomes CommandSavePreset (`urbex savepreset`):
instead of resolving a named preset, it now dumps the sender
dimension's live, resolved IDimensionInfo preset (post-inheritance,
post-overrides) to <gamedir>/urbex-export/<id path>.json via
PresetRE.CODEC, matching spec section 11.

Comments in Preset, Settings, PercentageSliderMath, SettingDescriptor
and SettingsCompletenessTest that named the deleted UrbexProfile/
Configuration classes are reworded to describe current reality
without depending on the removed types.
