+++ src/main/java/dev/krona/urbex/Urbex.java
@@ -1,13 +1,9 @@
 package dev.krona.urbex;
 
-import dev.krona.urbex.network.PacketRequestProfile;
-import dev.krona.urbex.network.PacketReturnProfileToClient;
 import dev.krona.urbex.setup.*;
 import dev.krona.urbex.varia.ServerAccess;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
-import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
-import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.fabricmc.loader.api.FabricLoader;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -39,8 +35,6 @@ public class Urbex implements ModInitializer {
         setup.preInit();
         setup.init();
 
-        registerNetworking();
-
         // Headless worldgen regression check; no-op without -Durbex.digestCheck
         DigestCheck.registerIfRequested();
 
@@ -48,12 +42,6 @@ public class Urbex implements ModInitializer {
         ServerLifecycleEvents.SERVER_STARTING.register(server -> ServerAccess.setServer(server));
         ServerLifecycleEvents.SERVER_STOPPED.register(server -> ServerAccess.setServer(null));
 
-    }
-
-    private void registerNetworking() {
-        PayloadTypeRegistry.clientboundPlay().register(PacketReturnProfileToClient.TYPE, PacketReturnProfileToClient.CODEC);
-        PayloadTypeRegistry.serverboundPlay().register(PacketRequestProfile.TYPE, PacketRequestProfile.CODEC);
-        ServerPlayNetworking.registerGlobalReceiver(PacketRequestProfile.TYPE, (payload, context) -> payload.handle());
     }
 
     public static Logger getLogger() {
