--- src/main/java/dev/krona/urbex/config/UrbexProfile.java
@@ -1,628 +1,0 @@
-package dev.krona.urbex.config;
-
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
-import dev.krona.urbex.Urbex;
-import dev.krona.urbex.setup.ModSetup;
-import net.minecraft.core.Holder;
-import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.network.FriendlyByteBuf;
-import net.minecraft.resources.Identifier;
-import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.state.BlockState;
-
-import java.util.Arrays;
-import java.util.Optional;
-
-public class UrbexProfile {
-
-    public static final String CATEGORY_CITY_ID = "lostcity";
-    public static final String CATEGORY_EXPLOSIONS = "explosions";
-    public static final String CATEGORY_CITIES = "cities";
-    public static final String CATEGORY_CLIENT = "client";
-
-    private final String name;
-    private final boolean isPublic;
-
-    private String description = "Default generation, common cities, explosions";
-    private String extraDescription = "";
-    private String warning = "";
-    private String worldStyle = "standard";
-    private String iconFile = "";
-    private Identifier icon;
-
-    public int DEBRIS_TO_NEARBYCHUNK_FACTOR = 200;
-
-    private String LIQUID_BLOCK = "minecraft:water";
-    private BlockState liquidBlock = null;
-
-    private String BASE_BLOCK = "minecraft:stone";
-    private BlockState baseBlock = null;
-
-    public float CHANCE_OF_RANDOM_LEAFBLOCKS = .1f;
-    public int THICKNESS_OF_RANDOM_LEAFBLOCKS = 2;
-    public boolean AVOID_FOLIAGE = false;
-
-    public float SCATTERED_CHANCE_MULTIPLIER = 1.0f;
-
-    public boolean RUBBLELAYER = true;
-    public float RUBBLE_DIRT_SCALE = 3.0f;
-    public float RUBBLE_LEAVE_SCALE = 6.0f;
-
-    public float RUIN_CHANCE = 0.05f;
-    public float RUIN_MINLEVEL_PERCENT = 0.8f;
-    public float RUIN_MAXLEVEL_PERCENT = 1.0f;
-
-    public int GROUNDLEVEL = 71;
-    public int SEALEVEL = -1;   // If -1 just use default
-
-    public boolean HIGHWAY_REQUIRES_TWO_CITIES = true;
-    public int HIGHWAY_LEVEL_FROM_CITIES_MODE = 0;
-    public float HIGHWAY_MAINPERLIN_SCALE = 50.0f;
-    public float HIGHWAY_SECONDARYPERLIN_SCALE = 10.0f;
-    public float HIGHWAY_PERLIN_FACTOR = 2.0f;
-    public int HIGHWAY_DISTANCE_MASK = 7;
-    public boolean HIGHWAY_SUPPORTS = true;
-
-    public float RAILWAY_DUNGEON_CHANCE = .01f;
-    public boolean RAILWAYS_CAN_END = false;
-    public boolean RAILWAYS_ENABLED = true;
-    public boolean RAILWAY_STATIONS_ENABLED = true;
-    public boolean RAILWAY_SURFACE_STATIONS_ENABLED = true;
-
-    public boolean EXPLOSIONS_IN_CITIES_ONLY = true;
-
-    public boolean EDITMODE = false;
-
-    public boolean GENERATE_NETHER = false;
-    public boolean GENERATE_SPAWNERS = true;
-    public float LIGHTING_DENSITY = 0.15f;
-    public float LOOT_DENSITY = 0.65f;
-    public boolean AVOID_WATER = false;
-
-    public float EXPLOSION_CHANCE = .002f;
-    public int EXPLOSION_MINRADIUS = 15;
-    public int EXPLOSION_MAXRADIUS = 35;
-    public int EXPLOSION_MINHEIGHT = 75;
-    public int EXPLOSION_MAXHEIGHT = 90;
-
-    public float MINI_EXPLOSION_CHANCE = .03f;
-    public int MINI_EXPLOSION_MINRADIUS = 5;
-    public int MINI_EXPLOSION_MAXRADIUS = 12;
-    public int MINI_EXPLOSION_MINHEIGHT = 60;
-    public int MINI_EXPLOSION_MAXHEIGHT = 100;
-
-    public double CITY_CHANCE = .01;    // If this is -1 we use perlin noise for rarity
-    public int CITY_MINRADIUS = 50;     // Only used for cityChance type
-    public int CITY_MAXRADIUS = 128;    // Only used for cityChance type
-
-    public double CITY_PERLIN_SCALE = 3;       // Used if CITY_CHANCE == -1
-    public double CITY_PERLIN_INNERSCALE = .1; // Used if CITY_CHANCE == -1
-    public double CITY_PERLIN_OFFSET = .1;     // Used if CITY_CHANCE == -1
-
-    // This threshold is used for both the cityChance variation as the perlin variation
-    public float CITY_THRESHOLD = .2f;
-
-    // These four values are used to modify the city factor based on distance from spawn
-    // If CITY_SPAWN_DISTANCE2 is 0 then this is disabled
-    // The city factor will be multiplied with CITY_SPAWN_MULTIPLIER1 at CITY_SPAWN_DISTANCE1
-    // The city factor will be multiplied with CITY_SPAWN_MULTIPLIER2 at CITY_SPAWN_DISTANCE2
-    // In between these two distances the factor is linearly interpolated
-    // Closer to spawn then CITY_SPAWN_DISTANCE1, CITY_SPAWN_MULTIPLIER1 is used
-    // Further from spawn then CITY_SPAWN_DISTANCE2, CITY_SPAWN_MULTIPLIER2 is used
-    public int CITY_SPAWN_DISTANCE1 = 0;
-    public int CITY_SPAWN_DISTANCE2 = 0;
-    public double CITY_SPAWN_MULTIPLIER1 = 1.0;
-    public double CITY_SPAWN_MULTIPLIER2 = 1.0;
-
-    // This threshold is used to select another citystyle if the city factor goes below this
-    public float CITY_STYLE_THRESHOLD = -1f;
-    public String CITY_STYLE_ALTERNATIVE = "";
-
-    public boolean CITY_AVOID_VOID = true;
-
-    public int CITY_LEVEL0_HEIGHT = 75;
-    public int CITY_LEVEL1_HEIGHT = 83;
-    public int CITY_LEVEL2_HEIGHT = 91;
-    public int CITY_LEVEL3_HEIGHT = 99;
-    public int CITY_LEVEL4_HEIGHT = 107;
-    public int CITY_LEVEL5_HEIGHT = 115;
-    public int CITY_LEVEL6_HEIGHT = 123;
-    public int CITY_LEVEL7_HEIGHT = 131;
-    public int CITY_MINHEIGHT = 50;
-    public int CITY_MAXHEIGHT = 150;
-
-    public int OCEAN_CORRECTION_BORDER = 4;
-
-    public int TERRAIN_FIX_LOWER_MIN_OFFSET = -4;
-    public int TERRAIN_FIX_LOWER_MAX_OFFSET = -3;
-    public int TERRAIN_FIX_UPPER_MIN_OFFSET = -1;
-    public int TERRAIN_FIX_UPPER_MAX_OFFSET = 1;
-
-    public float BUILDING_CHANCE = .3f;
-    public int BUILDING_MINFLOORS = 0;
-    public int BUILDING_MAXFLOORS = 8;
-    public int BUILDING_MINFLOORS_CHANCE = 4;
-    public int BUILDING_MAXFLOORS_CHANCE = 6;
-    public int BUILDING_MINCELLARS = 0;
-    public int BUILDING_MAXCELLARS = 3;
-    public float BUILDING_DOORWAYCHANCE = .6f;
-    public float BUILDING_FRONTCHANCE = .2f;
-
-    public int PRIMARY_ROAD_SPACING_X = 8;
-    public int PRIMARY_ROAD_SPACING_Z = 8;
-    public float PRIMARY_ROAD_OPTIONAL_CHANCE = .45f;
-    public int PRIMARY_ROAD_FORCE_EVERY = 4;
-    public int SECONDARY_ROAD_MIN_COUNT_X = 0;
-    public int SECONDARY_ROAD_MAX_COUNT_X = 2;
-    public int SECONDARY_ROAD_MIN_COUNT_Z = 0;
-    public int SECONDARY_ROAD_MAX_COUNT_Z = 2;
-    public int MINIMUM_ROAD_SEPARATION = 4;
-    public int MINIMUM_ROAD_EDGE_DISTANCE = 3;
-    public float TERTIARY_ROAD_CHANCE = .40f;
-    public int TERTIARY_ROAD_MIN_LENGTH = 2;
-    public int TERTIARY_ROAD_MAX_LENGTH = 5;
-    public float PLANNED_PRIMARY_BRIDGE_CHANCE = 1.0f;
-    public int PLANNED_PRIMARY_BRIDGE_MAX_LENGTH = 12;
-    public float OPEN_LOT_PARK_CHANCE = .8f;
-    public MultiBuildingStreetConflict MULTI_BUILDING_STREET_CONFLICT = MultiBuildingStreetConflict.OVERRIDE_MINOR;
-
-    public float CORRIDOR_CHANCE = .7f;
-    public float BRIDGE_CHANCE = .7f;
-    public float FOUNTAIN_CHANCE = .05f;
-
-    public boolean BRIDGE_SUPPORTS = true;
-    public boolean PARK_ELEVATION = true;
-    public boolean PARK_BORDER = true;
-    public int PARK_STREET_THRESHOLD = 3;
-
-    public boolean MULTI_USE_CORNER = false;
-    public boolean USE_AVG_HEIGHTMAP = false;
-
-    public int BEDROCK_LAYER = 1;
-
-    public float HORIZON = -1f;
-    public float FOG_RED = -1.0f;
-    public float FOG_GREEN = -1.0f;
-    public float FOG_BLUE = -1.0f;
-    public float FOG_DENSITY = -1.0f;
-
-    public String SPAWN_BIOME = "";
-    public String SPAWN_CITY = "";
-    public boolean SPAWN_NOT_IN_BUILDING = false;
-    public boolean FORCE_SPAWN_IN_BUILDING = false;
-    public String[] FORCE_SPAWN_BUILDINGS = new String[0];
-    public String[] FORCE_SPAWN_PARTS = new String[0];
-    public int SPAWN_CHECK_RADIUS = 200;
-    public int SPAWN_RADIUS_INCREASE = 100;
-    public int SPAWN_CHECK_ATTEMPTS = 20000;
-
-    public LandscapeType LANDSCAPE_TYPE = LandscapeType.DEFAULT;
-
-    public UrbexProfile(String name, boolean isPublic) {
-        this.name = name;
-        this.isPublic = isPublic;
-    }
-
-    public UrbexProfile(String name, String json) {
-        this.name = name;
-        Configuration config = new Configuration();
-        JsonElement element = JsonParser.parseString(json);
-        JsonObject root = element.getAsJsonObject();
-        config.fromJson(root);
-        this.isPublic = !root.has("public") || root.getAsJsonPrimitive("public").getAsBoolean();
-        init(config);
-    }
-
-    public void setIconFile(String iconFile) {
-        this.iconFile = iconFile;
-    }
-
-    public Identifier getIcon() {
-        if (icon != null) {
-            return icon;
-        }
-        if (iconFile == null || iconFile.isEmpty()) {
-            return null;
-        }
-        icon = Identifier.fromNamespaceAndPath(Urbex.MODID, iconFile);
-        return icon;
-    }
-
-    public boolean isPublic() {
-        return isPublic;
-    }
-
-    public void init(Configuration cfg) {
-        cfg.addCustomCategoryComment(UrbexProfile.CATEGORY_CITY_ID, "Settings related to Urbex city generation for the " + name + " profile");
-        cfg.addCustomCategoryComment(UrbexProfile.CATEGORY_EXPLOSIONS, "Settings related to explosions and damage for the " + name + " profile");
-        cfg.addCustomCategoryComment(UrbexProfile.CATEGORY_CITIES, "Settings related to city generation for the " + name + " profile");
-        cfg.addCustomCategoryComment(UrbexProfile.CATEGORY_CLIENT, "Client side settings for the " + name + " profile");
-
-        initLostcity(cfg);
-        initExplosions(cfg);
-        initCities(cfg);
-        initClient(cfg);
-    }
-
-    public String getCategoryLostcity() {
-        return UrbexProfile.CATEGORY_CITY_ID;
-    }
-
-    private void initClient(Configuration cfg) {
-        HORIZON = cfg.getFloat("horizon", UrbexProfile.CATEGORY_CLIENT, HORIZON, -1f, 256f, "This is used client-side (but only if the client has this mod) to set the height of the horizon");
-        FOG_RED = cfg.getFloat("fogRed", UrbexProfile.CATEGORY_CLIENT, FOG_RED, -1f, 1f, "This is used client-side (but only if the client has this mod) for the fog color");
-        FOG_GREEN = cfg.getFloat("fogGreen", UrbexProfile.CATEGORY_CLIENT, FOG_GREEN, -1f, 1f, "This is used client-side (but only if the client has this mod) for the fog color");
-        FOG_BLUE = cfg.getFloat("fogBlue", UrbexProfile.CATEGORY_CLIENT, FOG_BLUE, -1f, 1f, "This is used client-side (but only if the client has this mod) for the fog color");
-        FOG_DENSITY = cfg.getFloat("fogDensity", UrbexProfile.CATEGORY_CLIENT, FOG_DENSITY, -1f, 1f, "This is used client-side (but only if the client has this mod) for the fog density");
-    }
-
-    private void initLostcity(Configuration cfg) {
-        description = cfg.getString("description", UrbexProfile.CATEGORY_CITY_ID, description, "The description of this profile");
-        extraDescription = cfg.getString("extraDescription", UrbexProfile.CATEGORY_CITY_ID, extraDescription, "Additional information");
-        warning = cfg.getString("warning", UrbexProfile.CATEGORY_CITY_ID, warning, "Warning!");
-        worldStyle = cfg.getString("worldStyle", UrbexProfile.CATEGORY_CITY_ID, worldStyle, "The worldstyle used by this profile (defined in the assets)");
-        iconFile = cfg.getString("icon", UrbexProfile.CATEGORY_CITY_ID, iconFile, "The icon to use in the configuration screen (64x64)");
-
-        LIQUID_BLOCK = cfg.getString("liquidBlock", UrbexProfile.CATEGORY_CITY_ID, LIQUID_BLOCK, "Block to use as a liquid");
-        BASE_BLOCK = cfg.getString("baseBlock", UrbexProfile.CATEGORY_CITY_ID, BASE_BLOCK, "Block to use as the worldgen base");
-
-        SPAWN_BIOME = cfg.getString("spawnBiome", UrbexProfile.CATEGORY_CITY_ID, SPAWN_BIOME, "When this is set the player will always spawn in the given biome");
-        SPAWN_CITY = cfg.getString("spawnCity", UrbexProfile.CATEGORY_CITY_ID, SPAWN_CITY, "When this is set the player will always spawn in the given predefined city");
-        SPAWN_NOT_IN_BUILDING = cfg.getBoolean("spawnNotInBuilding", UrbexProfile.CATEGORY_CITY_ID, SPAWN_NOT_IN_BUILDING, "If this is true the player will not spawn in a building. This can be used in combination with the other spawn settings");
-        FORCE_SPAWN_IN_BUILDING = cfg.getBoolean("forceSpawnInBuilding", UrbexProfile.CATEGORY_CITY_ID, FORCE_SPAWN_IN_BUILDING, "If this is true the player will spawn in a building. This can be used in combination with the other spawn settings");
-        FORCE_SPAWN_BUILDINGS = cfg.getStringList("forceSpawnBuildings", UrbexProfile.CATEGORY_CITY_ID, FORCE_SPAWN_BUILDINGS, "A list of buildings that the player will spawn in if FORCE_SPAWN_IN_BUILDING is true. If this is empty then any building will do. This can be used in combination with 'forceSpawnParts'");
-        FORCE_SPAWN_PARTS = cfg.getStringList("forceSpawnParts", UrbexProfile.CATEGORY_CITY_ID, FORCE_SPAWN_PARTS, "A list of parts that the player will spawn in if FORCE_SPAWN_IN_BUILDING is true. If this is empty then any part will do. This can be used in combination with 'forceSpawnBuildings'");
-        SPAWN_CHECK_RADIUS = cfg.getInt("spawnCheckRadius", UrbexProfile.CATEGORY_CITY_ID, SPAWN_CHECK_RADIUS, 1, 100000, "The start radius to check for a valid spawn location. This value will be increased with 100 if it couldn't find a valid spot");
-        SPAWN_RADIUS_INCREASE = cfg.getInt("spawnRadiusIncrease", UrbexProfile.CATEGORY_CITY_ID, SPAWN_RADIUS_INCREASE, 1, 100000, "The radius increase to use if the spawn check radius is not large enough to find a valid spawn location");
-        SPAWN_CHECK_ATTEMPTS = cfg.getInt("spawnCheckAttempts", UrbexProfile.CATEGORY_CITY_ID, SPAWN_CHECK_ATTEMPTS, 1, 1000000, "The number of attempts to find a valid spawn location. If this is not enough then this will error");
-
-        SCATTERED_CHANCE_MULTIPLIER = cfg.getFloat("scatteredChanceMultiplier", UrbexProfile.CATEGORY_CITY_ID, SCATTERED_CHANCE_MULTIPLIER, 0.0f, 100.0f, "Multiplier for the chance a scattered building will generate. With 0 all scattered buildings are disabled");
-
-        TERRAIN_FIX_LOWER_MIN_OFFSET = cfg.getInt("terrainFixLowerMinOffset", UrbexProfile.CATEGORY_CITY_ID, TERRAIN_FIX_LOWER_MIN_OFFSET, -40, 40, "To fix terrain adjacent to cities, this is the minimum offset (relative to city base level) for the lower mesh used to potentially raise the terrain");
-        TERRAIN_FIX_LOWER_MAX_OFFSET = cfg.getInt("terrainFixLowerMaxOffset", UrbexProfile.CATEGORY_CITY_ID, TERRAIN_FIX_LOWER_MAX_OFFSET, -40, 40, "To fix terrain adjacent to cities, this is the maximum offset (relative to city base level) for the lower mesh used to potentially raise the terrain");
-        TERRAIN_FIX_UPPER_MIN_OFFSET = cfg.getInt("terrainFixUpperMinOffset", UrbexProfile.CATEGORY_CITY_ID, TERRAIN_FIX_UPPER_MIN_OFFSET, -40, 40, "To fix terrain adjacent to cities, this is the minimum offset (relative to city base level) for the upper mesh used to potentially lower the terrain");
-        TERRAIN_FIX_UPPER_MAX_OFFSET = cfg.getInt("terrainFixUpperMaxOffset", UrbexProfile.CATEGORY_CITY_ID, TERRAIN_FIX_UPPER_MAX_OFFSET, -40, 40, "To fix terrain adjacent to cities, this is the maximum offset (relative to city base level) for the upper mesh used to potentially lower the terrain");
-
-        CHANCE_OF_RANDOM_LEAFBLOCKS = cfg.getFloat("randomLeafBlockChance", UrbexProfile.CATEGORY_CITY_ID, CHANCE_OF_RANDOM_LEAFBLOCKS, 0.0f, 1.0f, "Chance that leafblocks will be generated at the border of a building and a street");
-        THICKNESS_OF_RANDOM_LEAFBLOCKS = cfg.getInt("randomLeafBlockThickness", UrbexProfile.CATEGORY_CITY_ID, THICKNESS_OF_RANDOM_LEAFBLOCKS, 1, 8,
-                "Frequency of leafblocks as seen from the sides of buildings");
-        AVOID_FOLIAGE = cfg.getBoolean("avoidFoliage", UrbexProfile.CATEGORY_CITY_ID, AVOID_FOLIAGE,
-                "If this is true then parks will have no foliage (trees and flowers currently)");
-
-        String[] landscapeNames = Arrays.stream(LandscapeType.values())
-                .map(LandscapeType::getName)
-                .toArray(String[]::new);
-        String type = cfg.getString("landscapeType", UrbexProfile.CATEGORY_CITY_ID,
-                LANDSCAPE_TYPE.getName(), "Type of landscape", landscapeNames);
-        LANDSCAPE_TYPE = LandscapeType.getTypeByName(type);
-        if (LANDSCAPE_TYPE == null) {
-            throw new RuntimeException("Bad landscape type: " + type + "!");
-        }
-
-        RUBBLELAYER = cfg.getBoolean("rubbleLayer", UrbexProfile.CATEGORY_CITY_ID, RUBBLELAYER,
-                "If this is true an alternative way to generate dirt/stone/sand + leave blocks is used that makes the city appear more overgrown");
-        RUBBLE_DIRT_SCALE = cfg.getFloat("rubbleDirtScale", UrbexProfile.CATEGORY_CITY_ID, RUBBLE_DIRT_SCALE, 0.0f, 100.0f,
-                "The scale of the dirt layer. Smaller values make the layer larger. Use 0 to disable");
-        RUBBLE_LEAVE_SCALE = cfg.getFloat("rubbleLeaveScale", UrbexProfile.CATEGORY_CITY_ID, RUBBLE_LEAVE_SCALE, 0.0f, 100.0f,
-                "The scale of the leave layer. Smaller values make the layer larger. Use 0 to disable");
-
-        RUIN_CHANCE = cfg.getFloat("ruinChance", UrbexProfile.CATEGORY_CITY_ID, RUIN_CHANCE, 0.0f, 1.0f,
-                "This gives the chance that a building is ruined");
-        RUIN_MINLEVEL_PERCENT = cfg.getFloat("ruinMinlevelPercent", UrbexProfile.CATEGORY_CITY_ID, RUIN_MINLEVEL_PERCENT, 0.0f, 1.0f,
-                "If a building is ruined this indicates the minimum start height for the ruin destruction layer");
-        RUIN_MAXLEVEL_PERCENT = cfg.getFloat("ruinMaxlevelPercent", UrbexProfile.CATEGORY_CITY_ID, RUIN_MAXLEVEL_PERCENT, 0.0f, 1.0f,
-                "If a building is ruined this indicates the maximum start height for the ruin destruction layer");
-
-        GROUNDLEVEL = cfg.getInt("groundLevel", UrbexProfile.CATEGORY_CITY_ID, GROUNDLEVEL, 2, 256, "Ground level");
-        SEALEVEL = cfg.getInt("seaLevel", UrbexProfile.CATEGORY_CITY_ID, SEALEVEL, -1, 256, "Sea level (-1 is default)");
-
-        initDensities(cfg);
-        BUILDING_CHANCE = cfg.getFloat("buildingChance", UrbexProfile.CATEGORY_CITY_ID, BUILDING_CHANCE, 0.0f, 1.0f, "The chance that a chunk in a city will have a building. Otherwise it will be a street");
-        BUILDING_MINFLOORS = cfg.getInt("buildingMinFloors", UrbexProfile.CATEGORY_CITY_ID, BUILDING_MINFLOORS, 0, 60, "The minimum number of floors (above ground) for a building (0 means the first floor only)");
-        BUILDING_MAXFLOORS = cfg.getInt("buildingMaxFloors", UrbexProfile.CATEGORY_CITY_ID, BUILDING_MAXFLOORS, 0, 60, "A cap for the amount of floors a city can have (above ground)");
-        BUILDING_MINFLOORS_CHANCE = cfg.getInt("buildingMinFloorsChance", UrbexProfile.CATEGORY_CITY_ID, BUILDING_MINFLOORS_CHANCE, 1, 60, "The amount of floors of a building is equal to: " +
-                "MINFLOORS + random(MINFLOORS_CHANCE + (cityFactor + .1f) * (MAXFLOORS_CHANCE - MINFLOORS_CHANCE))");
-        BUILDING_MAXFLOORS_CHANCE = cfg.getInt("buildingMaxFloorsChance", UrbexProfile.CATEGORY_CITY_ID, BUILDING_MAXFLOORS_CHANCE, 1, 60, "The amount of floors of a building is equal to: " +
-                "MINFLOORS + random(MINFLOORS_CHANCE + (cityFactor + .1f) * (MAXFLOORS_CHANCE - MINFLOORS_CHANCE))");
-
-        BUILDING_MINCELLARS = cfg.getInt("buildingMinCellars", UrbexProfile.CATEGORY_CITY_ID, BUILDING_MINCELLARS, 0, 20, "The minimum number of cellars (below ground). 0 means no cellar");
-        BUILDING_MAXCELLARS = cfg.getInt("buildingMaxCellars", UrbexProfile.CATEGORY_CITY_ID, BUILDING_MAXCELLARS, 0, 20, "The maximum number of cellars (below ground). 0 means no cellar");
-        BUILDING_DOORWAYCHANCE = cfg.getFloat("buildingDoorwayChance", UrbexProfile.CATEGORY_CITY_ID, BUILDING_DOORWAYCHANCE, 0.0f, 1.0f, "The chance that a doorway will be generated at a side of a building (on any level). Only when possible");
-        BUILDING_FRONTCHANCE = cfg.getFloat("buildingFrontChance", UrbexProfile.CATEGORY_CITY_ID, BUILDING_FRONTCHANCE, 0.0f, 1.0f, "The chance that a building will have a 'front' part if this is possible (i.e. adjacent street)");
-
-        PRIMARY_ROAD_SPACING_X = cfg.getInt("primaryRoadSpacingX", UrbexProfile.CATEGORY_CITY_ID, PRIMARY_ROAD_SPACING_X, 8, 128,
-                "Chunks between two candidate primary roads running north/south");
-        PRIMARY_ROAD_SPACING_Z = cfg.getInt("primaryRoadSpacingZ", UrbexProfile.CATEGORY_CITY_ID, PRIMARY_ROAD_SPACING_Z, 8, 128,
-                "Chunks between two candidate primary roads running east/west");
-        PRIMARY_ROAD_OPTIONAL_CHANCE = cfg.getFloat("primaryRoadOptionalChance", UrbexProfile.CATEGORY_CITY_ID, PRIMARY_ROAD_OPTIONAL_CHANCE, 0.0f, 1.0f,
-                "The chance that a primary road candidate that is not forced actually becomes a road");
-        PRIMARY_ROAD_FORCE_EVERY = cfg.getInt("primaryRoadForceEvery", UrbexProfile.CATEGORY_CITY_ID, PRIMARY_ROAD_FORCE_EVERY, 1, 16,
-                "Every Nth primary road candidate is always a road, which bounds how far apart two primary roads can be");
-        SECONDARY_ROAD_MIN_COUNT_X = cfg.getInt("secondaryRoadMinCountX", UrbexProfile.CATEGORY_CITY_ID, SECONDARY_ROAD_MIN_COUNT_X, 0, 128,
-                "The minimum number of secondary roads running north/south inside one primary block");
-        SECONDARY_ROAD_MAX_COUNT_X = cfg.getInt("secondaryRoadMaxCountX", UrbexProfile.CATEGORY_CITY_ID, SECONDARY_ROAD_MAX_COUNT_X, 0, 128,
-                "The maximum number of secondary roads running north/south inside one primary block");
-        SECONDARY_ROAD_MIN_COUNT_Z = cfg.getInt("secondaryRoadMinCountZ", UrbexProfile.CATEGORY_CITY_ID, SECONDARY_ROAD_MIN_COUNT_Z, 0, 128,
-                "The minimum number of secondary roads running east/west inside one primary block");
-        SECONDARY_ROAD_MAX_COUNT_Z = cfg.getInt("secondaryRoadMaxCountZ", UrbexProfile.CATEGORY_CITY_ID, SECONDARY_ROAD_MAX_COUNT_Z, 0, 128,
-                "The maximum number of secondary roads running east/west inside one primary block");
-        MINIMUM_ROAD_SEPARATION = cfg.getInt("minimumRoadSeparation", UrbexProfile.CATEGORY_CITY_ID, MINIMUM_ROAD_SEPARATION, 2, 32,
-                "The minimum number of chunks between two secondary roads on the same axis");
-        MINIMUM_ROAD_EDGE_DISTANCE = cfg.getInt("minimumRoadEdgeDistance", UrbexProfile.CATEGORY_CITY_ID, MINIMUM_ROAD_EDGE_DISTANCE, 2, 32,
-                "The minimum number of chunks between a secondary road and the primary road bounding its block");
-        TERTIARY_ROAD_CHANCE = cfg.getFloat("tertiaryRoadChance", UrbexProfile.CATEGORY_CITY_ID, TERTIARY_ROAD_CHANCE, 0.0f, 1.0f,
-                "The chance that a cell between roads receives a short dead-end access road");
-        TERTIARY_ROAD_MIN_LENGTH = cfg.getInt("tertiaryRoadMinLength", UrbexProfile.CATEGORY_CITY_ID, TERTIARY_ROAD_MIN_LENGTH, 1, 32,
-                "The minimum length in chunks of an access road");
-        TERTIARY_ROAD_MAX_LENGTH = cfg.getInt("tertiaryRoadMaxLength", UrbexProfile.CATEGORY_CITY_ID, TERTIARY_ROAD_MAX_LENGTH, 1, 32,
-                "The maximum length in chunks of an access road");
-        PLANNED_PRIMARY_BRIDGE_CHANCE = cfg.getFloat("plannedPrimaryBridgeChance", UrbexProfile.CATEGORY_CITY_ID, PLANNED_PRIMARY_BRIDGE_CHANCE, 0.0f, 1.0f,
-                "The chance that a primary road crossing open water is carried over it by a bridge");
-        PLANNED_PRIMARY_BRIDGE_MAX_LENGTH = cfg.getInt("plannedPrimaryBridgeMaxLength", UrbexProfile.CATEGORY_CITY_ID, PLANNED_PRIMARY_BRIDGE_MAX_LENGTH, 1, 64,
-                "The longest water crossing in chunks that a primary road will bridge");
-        OPEN_LOT_PARK_CHANCE = cfg.getFloat("openLotParkChance", UrbexProfile.CATEGORY_CITY_ID, OPEN_LOT_PARK_CHANCE, 0.0f, 1.0f,
-                "The chance that an open lot (a city chunk with neither a road nor a building) is furnished with a park part. Every open lot is grass regardless; this only decides whether a park part is placed on top of it");
-        MULTI_BUILDING_STREET_CONFLICT = MultiBuildingStreetConflict.byName(
-                cfg.getString("multiBuildingStreetConflict", UrbexProfile.CATEGORY_CITY_ID,
-                        MULTI_BUILDING_STREET_CONFLICT.name(),
-                        "How an accepted multi-building resolves against a planned road under its footprint"));
-
-        CORRIDOR_CHANCE = cfg.getFloat("corridorChance", UrbexProfile.CATEGORY_CITY_ID, CORRIDOR_CHANCE, 0.0f, 1.0f, "The chance that a chunk can possibly contain a corridor. " +
-                "There actually being a corridor also depends on the presence of adjacent corridors");
-        BRIDGE_CHANCE = cfg.getFloat("bridgeChance", UrbexProfile.CATEGORY_CITY_ID, BRIDGE_CHANCE, 0.0f, 1.0f, "The chance that a chunk can possibly contain a bridge. " +
-                "There actually being a bridge also depends on the presence of adjacent bridges and other conditions");
-        BRIDGE_SUPPORTS = cfg.getBoolean("bridgeSupports", UrbexProfile.CATEGORY_CITY_ID, BRIDGE_SUPPORTS,
-                "If true bridges get supports when needed. You can disable this if you have bridges that span void chunks");
-        PARK_ELEVATION = cfg.getBoolean("parkElevation", UrbexProfile.CATEGORY_CITY_ID, PARK_ELEVATION,
-                "If true parks get an extra layer of elevation");
-        PARK_BORDER = cfg.getBoolean("parkBorder", UrbexProfile.CATEGORY_CITY_ID, PARK_BORDER,
-                "If true parks border will use the street block as base.");
-        PARK_STREET_THRESHOLD = cfg.getInt("parkStreetThreshold", UrbexProfile.CATEGORY_CITY_ID, PARK_STREET_THRESHOLD,
-                0, 8, "Determines how many surrounding street it needs to generate a park, based on current chunk's park chance.");
-
-        FOUNTAIN_CHANCE = cfg.getFloat("fountainChance", UrbexProfile.CATEGORY_CITY_ID, FOUNTAIN_CHANCE, 0.0f, 1.0f, "The chance that a street section contains a fountain");
-
-        RAILWAY_DUNGEON_CHANCE = cfg.getFloat("railwayDungeonChance", UrbexProfile.CATEGORY_CITY_ID, RAILWAY_DUNGEON_CHANCE, 0.0f, 1.0f,
-                "The chance that a chunk next to a railway will have a railway dungeon");
-        RAILWAYS_CAN_END = cfg.getBoolean("railwaysCanEnd", UrbexProfile.CATEGORY_CITY_ID, RAILWAYS_CAN_END,
-                "If true the a place where a station would have been if there was a city above will have an 'ending' rail part if one side of the 'station' has no connections. Useful in case cities are rare");
-        RAILWAYS_ENABLED = cfg.getBoolean("railwaysEnabled", UrbexProfile.CATEGORY_CITY_ID, RAILWAYS_ENABLED,
-                "If true then railways are enabled. If false they are not (but stations will still generate)");
-        RAILWAY_STATIONS_ENABLED = cfg.getBoolean("railwayStationsEnabled", UrbexProfile.CATEGORY_CITY_ID, RAILWAY_STATIONS_ENABLED,
-                "If true then railway stations are enabled");
-        RAILWAY_SURFACE_STATIONS_ENABLED = cfg.getBoolean("railwaySurfaceStationsEnabled", UrbexProfile.CATEGORY_CITY_ID, RAILWAY_SURFACE_STATIONS_ENABLED,
-                "If true then railway surface station are enabled, otherwise there'll only be underground stations.");
-
-        MULTI_USE_CORNER = cfg.getBoolean("multiUseCorner", UrbexProfile.CATEGORY_CITY_ID, MULTI_USE_CORNER,
-                "Determine whether the multibuilding should use surrounding average level, or just top left corner level.");
-        USE_AVG_HEIGHTMAP = cfg.getBoolean("useAvgHeightmap", UrbexProfile.CATEGORY_CITY_ID, USE_AVG_HEIGHTMAP,
-                "City normal level generation is now dependent on surrounding heightmap and get an average of the surrounding heightmap. This uses extra performance since it needs to retrieve surrounding city heightmaps.");
-
-        HIGHWAY_REQUIRES_TWO_CITIES = cfg.getBoolean("highwayRequiresTwoCities", UrbexProfile.CATEGORY_CITY_ID, HIGHWAY_REQUIRES_TWO_CITIES,
-                "If true then a highway will only generate if both sides have a valid city. If false then one city is sufficient");
-        HIGHWAY_LEVEL_FROM_CITIES_MODE = cfg.getInt("highwayLevelFromCities", UrbexProfile.CATEGORY_CITY_ID, HIGHWAY_LEVEL_FROM_CITIES_MODE,
-                0, 3, "0 (take height from top-left city), 1 (take minimum height from both cities), 2 (take maximum height from both cities), 3 (take average height)");
-        HIGHWAY_DISTANCE_MASK = cfg.getInt("highwayDistanceMask", UrbexProfile.CATEGORY_CITY_ID, HIGHWAY_DISTANCE_MASK,
-                0, Integer.MAX_VALUE, "Mask to control how far highways can generate. Must be a power of 2 (minus 1). If 0 there are no highways at all");
-        HIGHWAY_MAINPERLIN_SCALE = cfg.getFloat("highwayMainPerlinScale", UrbexProfile.CATEGORY_CITY_ID, HIGHWAY_MAINPERLIN_SCALE, 1.0f, 1000.0f,
-                "For highways on a certain axis, this value is used to scale the perlin noise generator on the main axis. Increasing this value will increase the frequency of highways but make them smaller");
-        HIGHWAY_SECONDARYPERLIN_SCALE = cfg.getFloat("highwaySecondaryPerlinScale", UrbexProfile.CATEGORY_CITY_ID, HIGHWAY_SECONDARYPERLIN_SCALE, 1.0f, 1000.0f,
-                "For highways on a certain axis, this value is used to scale the perlin noise generator on the secondary axis. Increasing this value will increase the variation of nearby highways");
-        HIGHWAY_PERLIN_FACTOR = cfg.getFloat("highwayPerlinFactor", UrbexProfile.CATEGORY_CITY_ID, HIGHWAY_PERLIN_FACTOR, -100, 100,
-                "The highway perlin noise is compared to this value. Setting this to 0 would give 50% chance of a highway being at a spot. Note that highways only generate on chunks a multiple of 8. Setting this very high will prevent highways from generating");
-        HIGHWAY_SUPPORTS = cfg.getBoolean("highwaySupports", UrbexProfile.CATEGORY_CITY_ID, HIGHWAY_SUPPORTS,
-                "If true highways get supports when needed. You can disable this if you have highways that span void chunks");
-
-        BEDROCK_LAYER = cfg.getInt("bedrockLayer", UrbexProfile.CATEGORY_CITY_ID, BEDROCK_LAYER, 0, 10,
-                "The height of the bedrock layer that is generated at the bottom of some world types. Set to 0 to disable this and get default bedrock generation");
-
-        EDITMODE = cfg.getBoolean("editMode", UrbexProfile.CATEGORY_CITY_ID, EDITMODE,
-                "If true then this world is in edit mode");
-
-        GENERATE_NETHER = cfg.getBoolean("generateNether", UrbexProfile.CATEGORY_CITY_ID, GENERATE_NETHER,
-                "If true then generate a cavern type world in the Nether");
-        GENERATE_SPAWNERS = cfg.getBoolean("generateSpawners", UrbexProfile.CATEGORY_CITY_ID, GENERATE_SPAWNERS,
-                "If true then the buildings will be full of spawners");
-        AVOID_WATER = cfg.getBoolean("avoidWater", UrbexProfile.CATEGORY_CITY_ID, AVOID_WATER,
-                "If true then all water will be avoided (replaced with air)");
-    }
-
-    private void initDensities(Configuration cfg) {
-        if (!cfg.hasKey(CATEGORY_CITY_ID, "lightingDensity")
-                && cfg.hasKey(CATEGORY_CITY_ID, "generateLighting")) {
-            LIGHTING_DENSITY = cfg.getBoolean("generateLighting", CATEGORY_CITY_ID, false, "Legacy lighting switch")
-                    ? 1.0f : 0.0f;
-        }
-
-        boolean hasLegacyLoot = cfg.hasKey(CATEGORY_CITY_ID, "generateLoot")
-                || cfg.hasKey(CATEGORY_CITY_ID, "buildingWithoutLootChance")
-                || cfg.hasKey(CATEGORY_CITY_ID, "chestWithoutLootChance");
-        if (!cfg.hasKey(CATEGORY_CITY_ID, "lootDensity") && hasLegacyLoot) {
-            boolean enabled = cfg.getBoolean("generateLoot", CATEGORY_CITY_ID, true, "Legacy loot switch");
-            float buildingWithout = cfg.hasKey(CATEGORY_CITY_ID, "buildingWithoutLootChance")
-                    ? cfg.getFloat("buildingWithoutLootChance", CATEGORY_CITY_ID, 0.2f, 0.0f, 1.0f, "Legacy building exclusion")
-                    : 0.2f;
-            float chestWithout = cfg.hasKey(CATEGORY_CITY_ID, "chestWithoutLootChance")
-                    ? cfg.getFloat("chestWithoutLootChance", CATEGORY_CITY_ID, 0.2f, 0.0f, 1.0f, "Legacy chest exclusion")
-                    : 0.2f;
-            LOOT_DENSITY = enabled ? (1.0f - buildingWithout) * (1.0f - chestWithout) : 0.0f;
-        }
-
-        LIGHTING_DENSITY = clampDensity(cfg.getFloat("lightingDensity", CATEGORY_CITY_ID, LIGHTING_DENSITY,
-                0.0f, 1.0f, "Chance that an optional decorative-light marker places a light"));
-        LOOT_DENSITY = clampDensity(cfg.getFloat("lootDensity", CATEGORY_CITY_ID, LOOT_DENSITY,
-                0.0f, 1.0f, "Chance that a marked container receives a loot table"));
-    }
-
-    private static float clampDensity(float value) {
-        return Math.max(0.0f, Math.min(1.0f, value));
-    }
-
-    private void initCities(Configuration cfg) {
-        CITY_CHANCE = cfg.getDouble("cityChance", UrbexProfile.CATEGORY_CITIES, CITY_CHANCE, -1.0, 1.0, "The chance this chunk will be the center of a city (use -1 for perlin noise variant)");
-        CITY_SPAWN_DISTANCE1 = cfg.getInt("citySpawnDistance1", UrbexProfile.CATEGORY_CITIES, CITY_SPAWN_DISTANCE1, 0, 10000000, "At this distance from spawn (in blocks) the city factor will be equal to CITY_SPAWN_MULTIPLIER1");
-        CITY_SPAWN_DISTANCE2 = cfg.getInt("citySpawnDistance2", UrbexProfile.CATEGORY_CITIES, CITY_SPAWN_DISTANCE2, 0, 10000000, "At this distance from spawn (in blocks) the city factor will be equal to CITY_SPAWN_MULTIPLIER2. If this is 0 then no scaling is done");
-        CITY_SPAWN_MULTIPLIER1 = cfg.getDouble("citySpawnMultiplier1", UrbexProfile.CATEGORY_CITIES, CITY_SPAWN_MULTIPLIER1, 0, 1, "City factor at CITY_SPAWN_DISTANCE1");
-        CITY_SPAWN_MULTIPLIER2 = cfg.getDouble("citySpawnMultiplier2", UrbexProfile.CATEGORY_CITIES, CITY_SPAWN_MULTIPLIER2, 0, 1, "City factor at CITY_SPAWN_DISTANCE2");
-        CITY_MINRADIUS = cfg.getInt("cityMinRadius", UrbexProfile.CATEGORY_CITIES, CITY_MINRADIUS, 1, 2000, "The minimum radius of a city");
-        CITY_MAXRADIUS = cfg.getInt("cityMaxRadius", UrbexProfile.CATEGORY_CITIES, CITY_MAXRADIUS, 1, 2000, "The maximum radius of a city");
-        CITY_PERLIN_SCALE = cfg.getDouble("cityPerlinScale", UrbexProfile.CATEGORY_CITIES, CITY_PERLIN_SCALE, -1000000, 1000000, "The scale for the city rarity perlin map");
-        CITY_PERLIN_OFFSET = cfg.getDouble("cityPerlinOffset", UrbexProfile.CATEGORY_CITIES, CITY_PERLIN_OFFSET, -1000000, 1000000, "The offset for the city rarity perlin map");
-        CITY_PERLIN_INNERSCALE = cfg.getDouble("cityPerlinInnerScale", UrbexProfile.CATEGORY_CITIES, CITY_PERLIN_INNERSCALE, -1000000, 1000000, "The internal scale for the city rarity perlin map");
-        CITY_THRESHOLD = cfg.getFloat("cityThreshold", UrbexProfile.CATEGORY_CITIES, CITY_THRESHOLD, 0.0f, 1.0f, "The center and radius of a city define a sphere. " +
-                "This threshold indicates from which point a city is considered a city. " +
-                "This is important for calculating where cities are based on overlapping city circles (where the city thresholds are added)");
-        CITY_STYLE_THRESHOLD = cfg.getFloat("cityStyleThreshold", UrbexProfile.CATEGORY_CITIES, CITY_STYLE_THRESHOLD, 0.0f, 1.0f, "A city factor below this threshold will use the city style " +
-                "specified in 'cityStyleAlternative'");
-        CITY_STYLE_ALTERNATIVE = cfg.getString("cityStyleAlternative", UrbexProfile.CATEGORY_CITIES, CITY_STYLE_ALTERNATIVE, "Alternative city style. Used with cityStyleThreshold");
-        CITY_AVOID_VOID = cfg.getBoolean("cityAvoidVoid", UrbexProfile.CATEGORY_CITIES, CITY_AVOID_VOID, "Only used with floating landscape type: if true an additional detection is done to see if the chunk is void and in that case the city isn't generated there. Otherwise you might get city chunks on the border of islands which sometimes looks weird");
-
-        CITY_LEVEL0_HEIGHT = cfg.getInt("cityLevel0Height", UrbexProfile.CATEGORY_CITIES, CITY_LEVEL0_HEIGHT, 1, 384,
-                "Below this chunk height cities will be level 0");
-        CITY_LEVEL1_HEIGHT = cfg.getInt("cityLevel1Height", UrbexProfile.CATEGORY_CITIES, CITY_LEVEL1_HEIGHT, 1, 384,
-                "Below this chunk height cities will be level 1");
-        CITY_LEVEL2_HEIGHT = cfg.getInt("cityLevel2Height", UrbexProfile.CATEGORY_CITIES, CITY_LEVEL2_HEIGHT, 1, 384,
-                "Below this chunk height cities will be level 2");
-        CITY_LEVEL3_HEIGHT = cfg.getInt("cityLevel3Height", UrbexProfile.CATEGORY_CITIES, CITY_LEVEL3_HEIGHT, 1, 384,
-                "Below this chunk height cities will be level 3");
-        CITY_LEVEL4_HEIGHT = cfg.getInt("cityLevel4Height", UrbexProfile.CATEGORY_CITIES, CITY_LEVEL4_HEIGHT, 1, 384,
-                "Below this chunk height cities will be level 4");
-        CITY_LEVEL5_HEIGHT = cfg.getInt("cityLevel5Height", UrbexProfile.CATEGORY_CITIES, CITY_LEVEL5_HEIGHT, 1, 384,
-                "Below this chunk height cities will be level 5");
-        CITY_LEVEL6_HEIGHT = cfg.getInt("cityLevel6Height", UrbexProfile.CATEGORY_CITIES, CITY_LEVEL6_HEIGHT, 1, 384,
-                "Below this chunk height cities will be level 6");
-        CITY_LEVEL7_HEIGHT = cfg.getInt("cityLevel7Height", UrbexProfile.CATEGORY_CITIES, CITY_LEVEL7_HEIGHT, 1, 384,
-                "Below this chunk height cities will be level 7");
-        CITY_MINHEIGHT = cfg.getInt("cityMinHeight", UrbexProfile.CATEGORY_CITIES, CITY_MINHEIGHT, -1024, 2048,
-                "Below this height cities will not be generated");
-        CITY_MAXHEIGHT = cfg.getInt("cityMaxHeight", UrbexProfile.CATEGORY_CITIES, CITY_MAXHEIGHT, -1024, 2048,
-                "Above this height cities will not be generated");
-
-        OCEAN_CORRECTION_BORDER = cfg.getInt("oceanCorrectionBorder", UrbexProfile.CATEGORY_CITIES, OCEAN_CORRECTION_BORDER, -255, 255,
-                "Terrain correction offset that is used for chunks adjacent to city chunks that are in ocean biomes");
-    }
-
-    private void initExplosions(Configuration cfg) {
-        DEBRIS_TO_NEARBYCHUNK_FACTOR = cfg.getInt("debrisToNearbyChunkFactor", UrbexProfile.CATEGORY_EXPLOSIONS, DEBRIS_TO_NEARBYCHUNK_FACTOR, 1, 10000, "A factor that determines how much debris will overflow from nearby damaged chunks. Bigger numbers mean less debris");
-
-        EXPLOSION_CHANCE = cfg.getFloat("explosionChance", UrbexProfile.CATEGORY_EXPLOSIONS, EXPLOSION_CHANCE, 0.0f, 1.0f, "The chance that a chunk will contain an explosion");
-        EXPLOSION_MINRADIUS = cfg.getInt("explosionMinRadius", UrbexProfile.CATEGORY_EXPLOSIONS, EXPLOSION_MINRADIUS, 1, 1000, "The minimum radius of an explosion");
-        EXPLOSION_MAXRADIUS = cfg.getInt("explosionMaxRadius", UrbexProfile.CATEGORY_EXPLOSIONS, EXPLOSION_MAXRADIUS, 1, 3000, "The maximum radius of an explosion");
-        EXPLOSION_MINHEIGHT = cfg.getInt("explosionMinHeight", UrbexProfile.CATEGORY_EXPLOSIONS, EXPLOSION_MINHEIGHT, 1, 256, "The minimum height of an explosion");
-        EXPLOSION_MAXHEIGHT = cfg.getInt("explosionMaxHeight", UrbexProfile.CATEGORY_EXPLOSIONS, EXPLOSION_MAXHEIGHT, 1, 256, "The maximum height of an explosion");
-
-        MINI_EXPLOSION_CHANCE = cfg.getFloat("miniExplosionChance", UrbexProfile.CATEGORY_EXPLOSIONS, MINI_EXPLOSION_CHANCE, 0.0f, 1.0f, "The chance that a chunk will contain a mini explosion");
-        MINI_EXPLOSION_MINRADIUS = cfg.getInt("miniExplosionMinRadius", UrbexProfile.CATEGORY_EXPLOSIONS, MINI_EXPLOSION_MINRADIUS, 1, 1000, "The minimum radius of a mini explosion");
-        MINI_EXPLOSION_MAXRADIUS = cfg.getInt("miniExplosionMaxRadius", UrbexProfile.CATEGORY_EXPLOSIONS, MINI_EXPLOSION_MAXRADIUS, 1, 3000, "The maximum radius of a mini explosion");
-        MINI_EXPLOSION_MINHEIGHT = cfg.getInt("miniExplosionMinHeight", UrbexProfile.CATEGORY_EXPLOSIONS, MINI_EXPLOSION_MINHEIGHT, 1, 256, "The minimum height of a mini explosion");
-        MINI_EXPLOSION_MAXHEIGHT = cfg.getInt("miniExplosionMaxHeight", UrbexProfile.CATEGORY_EXPLOSIONS, MINI_EXPLOSION_MAXHEIGHT, 1, 256, "The maximum height of a mini explosion");
-
-        EXPLOSIONS_IN_CITIES_ONLY = cfg.getBoolean("explosionsInCitiesOnly", UrbexProfile.CATEGORY_EXPLOSIONS, EXPLOSIONS_IN_CITIES_ONLY,
-                "If this is true the center of an explosion can only be in a city (the blast can still affect non-city chunks)");
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public String getExtraDescription() {
-        return extraDescription;
-    }
-
-    public void setExtraDescription(String extraDescription) {
-        this.extraDescription = extraDescription;
-    }
-
-    public String getWarning() {
-        return warning;
-    }
-
-    public void setWarning(String warning) {
-        this.warning = warning;
-    }
-
-    public void setWorldStyle(String worldStyle) {
-        this.worldStyle = worldStyle;
-    }
-
-    public String getWorldStyle() {
-        return worldStyle;
-    }
-
-    public boolean isDefault() {
-        return LANDSCAPE_TYPE == LandscapeType.DEFAULT;
-    }
-
-    public boolean isFloating() {
-        return LANDSCAPE_TYPE == LandscapeType.FLOATING;
-    }
-
-    public boolean isCavern() { return LANDSCAPE_TYPE == LandscapeType.CAVERN; }
-
-
-    public BlockState getLiquidBlock() {
-        if (liquidBlock == null) {
-            Optional<Holder.Reference<Block>> b = BuiltInRegistries.BLOCK.get(Identifier.parse(LIQUID_BLOCK));
-            if (b.isEmpty()) {
-                ModSetup.getLogger().error("Bad liquid block: {}!", LIQUID_BLOCK);
-                liquidBlock = Blocks.WATER.defaultBlockState();
-            } else {
-                liquidBlock = b.get().value().defaultBlockState();
-            }
-        }
-        return liquidBlock;
-    }
-
-    public BlockState getBaseBlock() {
-        if (baseBlock == null) {
-            Optional<Holder.Reference<Block>>  b = BuiltInRegistries.BLOCK.get(Identifier.parse(BASE_BLOCK));
-            if (b.isEmpty()) {
-                ModSetup.getLogger().error("Bad base block: {}!", BASE_BLOCK);
-                baseBlock = Blocks.STONE.defaultBlockState();
-            } else {
-                baseBlock = b.get().value().defaultBlockState();
-            }
-        }
-        return baseBlock;
-    }
-
-    public void copyFrom(UrbexProfile other) {
-        Configuration config = new Configuration();
-        other.init(config);
-        init(config);
-    }
-
-    public JsonObject toJson(boolean readonly) {
-        Configuration config = new Configuration();
-        init(config);
-        JsonObject root = config.toJson(readonly);
-        root.addProperty("public", isPublic);
-        return root;
-    }
-
-    public void toBytes(FriendlyByteBuf buf) {
-        JsonObject jsonObject = toJson(false);
-        buf.writeUtf(jsonObject.toString());
-    }
-
-    public Configuration toConfiguration() {
-        Configuration config = new Configuration();
-        init(config);
-        return config;
-    }
-
-    public void copyFromConfiguration(Configuration config) {
-        init(config);
-    }
-}
