--- src/main/java/dev/krona/urbex/network/PacketReturnProfileToClient.java
@@ -1,33 +1,0 @@
-package dev.krona.urbex.network;
-
-import dev.krona.urbex.Urbex;
-import net.minecraft.core.registries.Registries;
-import net.minecraft.network.RegistryFriendlyByteBuf;
-import net.minecraft.network.codec.ByteBufCodecs;
-import net.minecraft.network.codec.StreamCodec;
-import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
-import net.minecraft.resources.ResourceKey;
-import net.minecraft.resources.Identifier;
-import net.minecraft.world.level.Level;
-
-public record PacketReturnProfileToClient(ResourceKey<Level> dimension, String profile) implements CustomPacketPayload {
-
-    public static final Identifier ID = Identifier.fromNamespaceAndPath(Urbex.MODID, "returnprofile");
-    public static final CustomPacketPayload.Type<PacketReturnProfileToClient> TYPE = new Type<>(ID);
-
-    public static final StreamCodec<RegistryFriendlyByteBuf, PacketReturnProfileToClient> CODEC = StreamCodec.composite(
-            ResourceKey.streamCodec(Registries.DIMENSION), PacketReturnProfileToClient::dimension,
-            ByteBufCodecs.STRING_UTF8, PacketReturnProfileToClient::profile,
-            PacketReturnProfileToClient::new);
-
-    @Override
-    public Type<? extends CustomPacketPayload> type() {
-        return TYPE;
-    }
-
-
-    public void handle() {
-        // @todo 1.14
-//            WorldTypeTools.setProfileFromServer(dimension, profile);
-    }
-}
