+++ src/main/java/dev/krona/urbex/setup/ClientSetup.java
@@ -1,14 +1,11 @@
 package dev.krona.urbex.setup;
 
-import dev.krona.urbex.network.PacketReturnProfileToClient;
 import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 
 public class ClientSetup implements ClientModInitializer {
 
     @Override
     public void onInitializeClient() {
         ClientEventHandlers.register();
-        ClientPlayNetworking.registerGlobalReceiver(PacketReturnProfileToClient.TYPE, (payload, context) -> payload.handle());
     }
 }
